[#18] refactor: ApiResponse 위치 변경
